secure_messenger/requirements.txt
flask
flask-wtf
cryptography
